Move to string namespace
